Project set 2 test passes:
+ Remember to actually add counted days to `countedDays` set (woops)
+ Tweak day type logic to include either first OR last project days as
full days when contiguous with other project.

Expect to need to change day type logic again on next set because the
overall first project day of a set, even if its project is contiguous
with the next one, needs to be counted as a travel day. Ditto for last
day.